# emas/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Job:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      deadline_status:
        $ref: '#/definitions/domain.JobDeadlineStatus'
      job_id:
        type: string
      notes:
        type: string
      priority:
        type: string
      product_id:
        type: string
      quantity_completed:
        type: integer
      quantity_total:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  domain.JobDeadlineStatus:
    properties:
      is_late:
        type: boolean
      late_by:
        description: 'human-readable: "2 days", "4 hours", "on time"'
        type: string
    type: object
...
  dto.CreateJobRequest:
    properties:
      deadline:
        description: RFC3339
        type: string
      notes:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      product_id:
        type: string
      quantity_total:
        type: integer
      slots:
        description: optional split slots
        items:
          $ref: '#/definitions/dto.CreateSlotRequest'
        type: array
    required:
    - product_id
    - quantity_total
    type: object
...
  dto.CreateSlotRequest:
    properties:
      allocation_percent:
        type: number
      batch_sequence:
        type: integer
      buffer_mins:
        type: integer
      cleaning_mins:
        type: integer
      duration_mins:
        type: integer
      is_parallel:
        type: boolean
      job_step_id:
        description: optional, for split
        type: string
      machine_id:
        type: string
      prep_mins:
        type: integer
      processing_mins:
        type: integer
      proposal_id:
        type: string
      quantity:
        type: integer
      split_group_id:
        type: string
      start_time:
        description: RFC3339
        type: string
    required:
    - duration_mins
    - machine_id
    - quantity
    - start_time
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UpdateJobRequest:
    properties:
      deadline:
        type: string
      notes:
        type: string
      priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      quantity_total:
        type: integer
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Factory operations API.
  title: eMas Factory API
  version: "1.0"
paths:
...
  /jobs:
    get:
      consumes:
      - application/json
      description: List jobs with filters
      parameters:
      - description: Filter by product
        in: query
        name: product_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by priority (low|medium|high|urgent)
        in: query
        name: priority
        type: string
      - description: Filter by machine
        in: query
        name: machine_id
        type: string
      - description: RFC3339 start
        in: query
        name: start
        type: string
      - description: RFC3339 end
        in: query
        name: end
        type: string
      - description: created_at|deadline|priority|quantity_total|completion
        in: query
        name: sort_by
        type: string
      - description: asc|desc
        in: query
        name: sort_dir
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Job'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List jobs
      tags:
      - job
    post:
      consumes:
      - application/json
      description: Create a job
      parameters:
      - description: Create Job Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Job'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Create a job
      tags:
      - job
  /jobs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a job and clear all slot assignments tied to this job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete a job
      tags:
      - job
    get:
      consumes:
      - application/json
      description: Get a job by ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Job'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get a job by ID
      tags:
      - job
    put:
      consumes:
      - application/json
      description: Update mutable job fields
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Job Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Job'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update a job
      tags:
      - job
...
swagger: "2.0"